backend:error handling subscrip_service

+++ backend/app/services/subscription_service.py
@@ -3,7 +3,7 @@ from repos.park_repo import ParkRepo
 from repos.user_repo import UserRepo
 from models.subscription import Subscription
 from representations.subscription import CreateSubscriptionRequest, SubscriptionSchema
-
+from resources.errors import BadRequestError
 
 class SubscriptionService:
     def __init__(self, subscription_repo: SubscriptionRepo, park_repo: ParkRepo, user_repo: UserRepo):
@@ -14,15 +14,16 @@ class SubscriptionService:
         self.subscriptions_schema = SubscriptionSchema(many=True)
 
     def create(self, create_subscription_request: CreateSubscriptionRequest):
+        subscription = Subscription()
         park, error = self.park_repo.get_by_id(create_subscription_request.park_id)
         if park is not None and error is None:
-            post.park = park
+            subscription.park = park
         else: 
             return None, error
         
-        user, error = self.user_repo.get_by_id(create_post_request.user_id)
-         if user is not None and error is None:
-            post.user = user
+        user, error = self.user_repo.get_by_id(create_subscription_request.user_id)
+        if user is not None and error is None:
+            subscription.user = user
         else:
             return None, error
         
@@ -35,5 +36,10 @@ class SubscriptionService:
         return self.subscriptions_schema.dump(subscriptions).data
 
     def delete_by_id(self, id: str):
-        subscription = self.subscription_repo.get_by_id(id)
-        self.subscription_repo.delete(subscription)
+        subscription, error = self.subscription_repo.get_by_id(id)
+        if park is not None and error is None:
+            return self.subscription_repo.delete(subscription)
+        elif park is None and error is not None:
+            return None, BadRequestError
+        else:
+            return None, error
